Using factory to create users and players in PlayerTest
