+++ database/factories/UserFactory.php
@@ -35,7 +35,8 @@ class UserFactory extends Factory
             'surname'           => $fakeSurname,
             'email'             => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
-            'password'          => static::$password ??= Hash::make('password'),
+          //'password'          => static::$password ??= Hash::make('password'),
+            'password'          => Hash::make(fake()->password()),
             'remember_token'    => Str::random(10),
         ];
     }
